requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
ecdsa==0.19.1
fastapi==0.115.12
frozenlist==1.6.0
geoip2==5.0.1
greenlet==3.2.1
h11==0.14.0
idna==3.10
maxminddb==2.6.3
multidict==6.4.3
passlib==1.7.4
propcache==0.3.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyMySQL==1.1.1
python-dotenv==1.1.0
python-jose==3.4.0
redis==5.2.1
-e git+ssh://git@github.com/lulucasalves/refund-backend.git@a0dd41fba5ce41267943bc2771d62defa00f8ab5#egg=refundBackend
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uuid==1.30
uvicorn==0.34.2
yarl==1.20.0
